Pressing mouse button inside a dragstart event causes a crash https://bugs.webkit.org/show_bug.cgi?id=115296
Reviewed by Darin Adler.

Source/WebCore:

Add a missing null pointer check. We should better encapsulate the states in DragState in the long term
but this is good enough for now.

Test: fast/events/mousedown-inside-dragstart-should-not-cause-crash.html

* page/EventHandler.cpp:
(WebCore::EventHandler::handleDrag):

LayoutTests:

Added a regression test. While the bug report involves opening inspector and setting a breakpoint,
a simpler reduction that uses eventSender significantly reduces the complexity.

* fast/events/mousedown-inside-dragstart-should-not-cause-crash-expected.txt: Added.
* fast/events/mousedown-inside-dragstart-should-not-cause-crash.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@149254 268f45cc-cd09-0410-ab3c-d52691b4dbfc