[Qt] REGRESSION (r148975) shape-inside-percentage.html and shape-inside-overflow-fixed-dimensions.html are failing after r148975 https://bugs.webkit.org/show_bug.cgi?id=115053
Reviewed by Alexandru Chiculita.

I changed the tests to use Ahem font (overflow will be the same on every port) and modify the markup in order to be the similar
for the two tests. The change fixes both tests and make them passing on Qt as well. I removed the tests from the TestExpectations.

* fast/exclusions/shape-inside/shape-inside-overflow-fixed-dimensions-expected.html:
* fast/exclusions/shape-inside/shape-inside-overflow-fixed-dimensions.html:
* fast/exclusions/shape-inside/shape-inside-percentage-expected.html:
* fast/exclusions/shape-inside/shape-inside-percentage.html:
* platform/qt/TestExpectations:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@149857 268f45cc-cd09-0410-ab3c-d52691b4dbfc