REGRESSION (r145680): No box shadow rendered on element with positioned child that obscures it https://bugs.webkit.org/show_bug.cgi?id=115840
Source/WebCore:

Reviewed by Antti Koivisto.

In r107836, we moved some box-shadow painting into the paintFillLayer code for performance.
However, in r145680 we started to skip background painting when we know the background is
obscured. This broke shadow painting in some cases.

Fix by always painting the background fill layers if we know that they will also
paint the shadow.

Test: fast/box-shadow/box-shadow-obscured-backgrounds.html

* rendering/RenderBox.cpp:
(WebCore::RenderBox::paintBackground):

LayoutTests:

Reviewed by Antti Koivisto.

* fast/box-shadow/box-shadow-obscured-backgrounds-expected.html: Added.
* fast/box-shadow/box-shadow-obscured-backgrounds.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@149918 268f45cc-cd09-0410-ab3c-d52691b4dbfc