Don't pass malloced pointers as out-of-line data when sending Mach messages https://bugs.webkit.org/show_bug.cgi?id=115970 <rdar://problem/13144680>
Reviewed by Sam Weinig.

Give ArgumentEncoder an inline buffer of 512 bytes. If we need more than that, use mmap to allocate data
since that gives us fresh zero-filled memory that we can safely share.

* Platform/CoreIPC/ArgumentEncoder.cpp:
(CoreIPC::ArgumentEncoder::ArgumentEncoder):
Set up the buffer pointers to point to the inline buffer.

(CoreIPC::ArgumentEncoder::~ArgumentEncoder):
Unmap the buffer if necessary.

(CoreIPC::ArgumentEncoder::grow):
Grow the allocation exponentially, rounded up to the nearest page. This is a simplification from the
current strategy, but most messages are either tiny in which case they will fit inside the inline buffer,
or big in which case we'll end up doing less allocations + memory copying.

* Platform/CoreIPC/ArgumentEncoder.h:
(ArgumentEncoder):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@149944 268f45cc-cd09-0410-ab3c-d52691b4dbfc