[Mac] captions menu is not positioned correctly in full-screen https://bugs.webkit.org/show_bug.cgi?id=116103
Update the positioning properties of the captions menu in full-screen
following the fix for https://webkit.org/b/115968.

Reviewed by Darin Adler.

* css/fullscreenQuickTime.css:
(video:-webkit-full-screen::-webkit-media-controls-closed-captions-container):
Add 50px to the bottom position since we're now positioned relative to the
media element, and use a calc() command to be right-aligned to the captions
icon in the media controller. The max-width and max-height properties follow
the same technique used for the non-full-screen mode.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@150075 268f45cc-cd09-0410-ab3c-d52691b4dbfc