Avoid backing store for the WebFrameView's layer when the WebView is layer-backed https://bugs.webkit.org/show_bug.cgi?id=116172
Reviewed by Tim Horton.

When the WebView is layer-backed, AppKit will create a layer with backing store for
the WebFrameView because it implements drawRect:. However, this method only paints
when there is no documentView, so this layer’s backing store is wasteful.

We can avoid allocation of this backing store by implementing -wantsUpdateLayer
and -updateLayer, and setting the view’s backgroundColor.

* WebView/WebFrameView.mm:
(-[WebFrameView wantsUpdateLayer]):
(-[WebFrameView updateLayer]):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@150168 268f45cc-cd09-0410-ab3c-d52691b4dbfc