[Cairo] Canvas-shadow behavior is not being as expected https://bugs.webkit.org/show_bug.cgi?id=108897
Patch by Rashmi Shyamasundar <rashmi.s2@samsung.com> on 2013-05-18
Reviewed by Martin Robinson.

Source/WebCore:

ShadowBlur::endShadowLayer copies the image from shadowContext to cairoContext.
CairoContext-path should be empty for doing this copy. Otherwise, the
original-image area will also get filled with the shadow.

Test: fast/canvas/canvas-image-shadow.html

* platform/graphics/cairo/GraphicsContextCairo.cpp:
(WebCore::drawPathShadow):

LayoutTests:

Test to verify the shadow of an image drawn on canvas.
This test uses an image whose size is smaller than,
the size of the rectangle which should be filled with the image.

* fast/canvas/canvas-image-shadow-expected.txt: Added.
* fast/canvas/canvas-image-shadow.html: Added.
* fast/canvas/script-tests/canvas-image-shadow.js: Added.
(draw):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@150370 268f45cc-cd09-0410-ab3c-d52691b4dbfc