[GTK] Pass GTEST_HAS_RTTI=0 cppflag to TestWebKitAPI programs at compilation-time https://bugs.webkit.org/show_bug.cgi?id=116450
Reviewed by Martin Robinson.

The -fno-rtti flag is used when compiling C++ source code throughout the project. The TestWebKitAPI
programs fail to compile because of this when using the C++11 standard as the GTest headers assume by
default that the RTTI support is available. This can be turned off by passing GTEST_HAS_RTTI=0 to the
TestWebKitAPI programs when compiling them.

* TestWebKitAPI/GNUmakefile.am:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@150420 268f45cc-cd09-0410-ab3c-d52691b4dbfc