Improve loader/go-back-cached-main-resource.html test https://bugs.webkit.org/show_bug.cgi?id=116709
Patch by Manuel Rego Casasnovas <rego@igalia.com> on 2013-05-24
Reviewed by Alexey Proskuryakov.

It is only needed to register the listeners on the initial load as they
are already present when restoring from page cache. Modified the test to
use event.persisted property to know if we are handling the first load
in order to register the listeners only once.

* loader/resources/first-page.html: Rename registerListeners() function
to a more accurate name didShowPage() and use event.persisted to
register the listeners only in the first load.
* loader/resources/other-page.html: Ditto.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@150645 268f45cc-cd09-0410-ab3c-d52691b4dbfc