[CSS Regions] Content that has overflow: scroll cannot be scrolled by dragging the scroll thumbs with the mouse https://bugs.webkit.org/show_bug.cgi?id=113703
Patch by Mihai Maerean <mmaerean@adobe.com> on 2013-05-29
Reviewed by Darin Adler.

Source/WebCore:

The point inside a box that's inside a region has its coordinates relative to the region, not the FlowThread
that is its container in the RenderObject tree.

Previously, the coordinates were computed relatively to the FlowThread which doesn't really have a location in
the page.

Tests: fast/regions/drag-scrollbars-of-content.html
       fast/regions/hover-in-second-region.html

* rendering/RenderBox.cpp:
(WebCore::RenderBox::offsetFromContainer):
* rendering/RenderBoxModelObject.cpp:
(WebCore::RenderBoxModelObject::mapAbsoluteToLocalPoint):

LayoutTests:

* fast/regions/drag-scrollbars-of-content-expected.html: Added.
* fast/regions/drag-scrollbars-of-content.html: Added.
* fast/regions/hover-in-second-region-expected.html: Added.
* fast/regions/hover-in-second-region.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@150881 268f45cc-cd09-0410-ab3c-d52691b4dbfc