Add a special case for SelectorDataList::execute when there is only one selector https://bugs.webkit.org/show_bug.cgi?id=116891
Patch by Benjamin Poulain <bpoulain@apple.com> on 2013-05-29
Reviewed by Andreas Kling.

In the vast majority of cases, querySelector is called with a single selector.

This patch takes that input into account to improve the performance of
SelectorDataList::execute(). For the case of a single selector, we avoid three
branches executed for every node of the traversal.

This improves QuerySelector's time by 2.3%.

* dom/SelectorQuery.cpp:
(WebCore::SelectorDataList::execute):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@150944 268f45cc-cd09-0410-ab3c-d52691b4dbfc