[GTK] Needs to check for harfbuzz-icu https://bugs.webkit.org/show_bug.cgi?id=116978
Patch by Alberto Garcia <agarcia@igalia.com> on 2013-05-30
Reviewed by Xan Lopez.

HarfBuzz 0.9.18 splits harbuzz-icu into a separate library so we
also need to check for it in order to get the necessary flags for
the compiler and the linker.

We keep this conditional for now since we still want to support
earlier versions of HarfBuzz.

* Source/autotools/FindDependencies.m4:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@150963 268f45cc-cd09-0410-ab3c-d52691b4dbfc