[CSSRegions] Respect renderer creation constraints when element is part of named flow https://bugs.webkit.org/show_bug.cgi?id=116790
Reviewed by David Hyatt.

Source/WebCore:

Before adding a child to a named flow, we have to make sure that the rules for
creation are properly checked.

Test: fast/regions/br-content-node-crash.xhtml

* rendering/RenderNamedFlowThread.cpp:
(WebCore::RenderNamedFlowThread::isChildAllowed):
Call isChildAllowed on the original parent renderer, before reparenting to
named flow.
* rendering/RenderNamedFlowThread.h:

LayoutTests:

The test needs to be xhtml so that br element, child of frame set, remains
child of the frame set when the document is parsed.

* fast/regions/br-content-node-crash-expected.txt: Added.
* fast/regions/br-content-node-crash.xhtml: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@151204 268f45cc-cd09-0410-ab3c-d52691b4dbfc