[GTK] Parameters 'inResult' and 'resolver' from function 'webkit_dom_document_evaluate' should be allowed to be NULL https://bugs.webkit.org/show_bug.cgi?id=117129
Patch by Diego Pino Garcia <dpino@igalia.com> on 2013-06-06
Reviewed by Xan Lopez.

Source/WebCore:

At this moment there was a temporary fix that allowed parameter
'inResult' to be NULL (see: webk.it/42115). However, there was no fix
for parameter 'resolver'.

This patch refactors the code of the previous fix, moving the code for
determine whether a parameter can be NULL or not to GetGReturnMacro.
The solution is quite general and will alow to add other parameters in
the future if needed.

* bindings/scripts/CodeGeneratorGObject.pm:
(GetGReturnMacro): Pass functionName, as in some cases the code
generated depends on the paramName and the functionName
(ParamCanBeNull): Checks if a parameter is allowed to be NULL
(GenerateFunction):

Source/WebKit/gtk:

Add test for function 'webkit_dom_document_evaluate'.

* tests/testdomdocument.c:
(test_dom_document_evaluate): Checks function dom_document_evaluate,
executes an XPath expression on a HTML document.
(main):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@151267 268f45cc-cd09-0410-ab3c-d52691b4dbfc