REGRESSION(r150633): Find on page non-focused text highlight color is bright yellow. <rdar://problem/14098882> <http://webkit.org/b/117371>
Reviewed by Anders Carlsson.

There was a mistake in the parameter ordering which still compiled because of
implicit bool/unsigned conversion.
Changed the interface of Page::findMatchesForText() to use enums instead of bools.

No test because the color of the highlighted matches is not available through APIs.

* page/Page.h:
* page/Page.cpp:
(WebCore::Page::findMatchesForText):
(WebCore::Page::markAllMatchesForText):
(WebCore::Page::countFindMatches):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@151404 268f45cc-cd09-0410-ab3c-d52691b4dbfc