[BlackBerry] Test editing/execCommand/indent-paragraphs.html times out https://bugs.webkit.org/show_bug.cgi?id=117235
Patch by Carlos Garcia Campos <cgarcia@igalia.com> on 2013-06-11
Reviewed by Rob Buis.

PR 346414.

After r150224 ResourceHandle::start() can fail returning false so
that the ResourceHandle is destroyed and the error not
handled. Handle the errors by scheduling a InvalidURLFailure and
returning true when startJob fails to make sure the ResourceHandle
is not destroyed and the error is handled and notified to the
client.

Fixes timeout in editing/execCommand/indent-paragraphs.html.

* platform/network/blackberry/ResourceHandleBlackBerry.cpp:
(WebCore::ResourceHandle::start):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@151426 268f45cc-cd09-0410-ab3c-d52691b4dbfc