REGRESSION (r148367): Facebook and Twitter icons at macworld.com are stacked vertically, obscuring Twitter one https://bugs.webkit.org/show_bug.cgi?id=117284
Reviewed by David Hyatt.

Source/WebCore:

Test: fast/text/whitespace/inline-whitespace-wrapping-7.html

We were looking for line breaks in trailing collapsed whitespace even when inside a nowrap inline.
It's wrong to do this as we don't know yet if we will want or need to break - that decision has to
be deferred until we enter a part of the line that is autowrap.

* rendering/RenderBlockLineLayout.cpp:
(WebCore::RenderBlock::LineBreaker::nextSegmentBreak):

LayoutTests:

* fast/text/whitespace/inline-whitespace-wrapping-7-expected.html: Added.
* fast/text/whitespace/inline-whitespace-wrapping-7.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@151613 268f45cc-cd09-0410-ab3c-d52691b4dbfc