[GStreamer] [texmap] the bytesPerLine is wrong when updating the texture https://bugs.webkit.org/show_bug.cgi?id=117551
Patch by Víctor Manuel Jáquez Leal <vjaquez@igalia.com> on 2013-06-19
Reviewed by Philippe Normand.

Source/WebCore:

The bytesPerLine (or stride) is wrong when updating the texture
content. This is because the naturalSize() is not updated to the
current buffer. This patch extracts the current value of the size and
the stride.

No tests required.

* platform/graphics/gstreamer/MediaPlayerPrivateGStreamerBase.cpp:
(WebCore::MediaPlayerPrivateGStreamerBase::updateTexture):
(WebCore::MediaPlayerPrivateGStreamerBase::paintToTextureMapper):

LayoutTests:

* platform/gtk-wk2/TestExpectations:
enabled media/video-display-aspect-ratio.html

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@151733 268f45cc-cd09-0410-ab3c-d52691b4dbfc