REGRESSION(r149652): accessing items in .children via id doesn't work when element is not rooted in DOM tree https://bugs.webkit.org/show_bug.cgi?id=117836
Reviewed by Benjamin Poulain.

Source/WebCore: 

When the root node of a HTML collection is not in the document or in a shadow tree,
we shouldn't use its tree scope's id and name maps to find name getters.

Always use the slow path in such cases.

Test: fast/dom/htmlallcollection-detached-node-children.html

* html/HTMLCollection.cpp:
(WebCore::HTMLCollection::namedItem):

LayoutTests: 

Add a regression test for named getter for a detached element.

* fast/dom/htmlallcollection-detached-node-children-expected.txt: Added.
* fast/dom/htmlallcollection-detached-node-children.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@151821 268f45cc-cd09-0410-ab3c-d52691b4dbfc