An element with -webkit-user-select: all should be selected on single click https://bugs.webkit.org/show_bug.cgi?id=117864
Reviewed by Darin Adler.

Source/WebCore: 

Set m_mouseDownMayStartSelect true even when the node cannot start selection
if the node was inside an element with -webkit-user-select: all.

Test: editing/selection/user-select-all-with-single-click.html

* page/EventHandler.cpp:
(WebCore::canMouseDownStartSelect):

LayoutTests: 

Add tests for single clicking on an element with user-select: all.
We select the entire element regardless of whether the user had clicked on text or an image
to be consistent even though Firefox requires a double click on the latter case.

* editing/selection/user-select-all-with-image-single-click-expected.txt: Added.
* editing/selection/user-select-all-with-image-single-click.html: Added.
* editing/selection/user-select-all-with-single-click-expected.txt: Added.
* editing/selection/user-select-all-with-single-click.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@152198 268f45cc-cd09-0410-ab3c-d52691b4dbfc